fix: keep pub/sub receiver alive to enable message delivery
The receiver from register_subscriber() must be kept alive in CommandContext,
otherwise the send channel closes and PUBLISH can't deliver messages.

Changes:
- Store pubsub_receiver in CommandContext (with #[allow(dead_code)])
- Update all constructors to keep receiver instead of dropping it
- Fix test expectation for mixed subscribe/psubscribe (counts sends, not clients)

Test results:
✅ All 19 Pub/Sub integration tests passing
✅ All 2,594 workspace tests passing
✅ No clippy warnings

This fixes CI failures on Ubuntu, macOS, and Windows.

+++ ferris-network/tests/pubsub_commands.rs
@@ -394,9 +394,9 @@ async fn test_mixed_subscribe_and_psubscribe() {
     // Should match both subscription and pattern
     let result = publisher.cmd(&["PUBLISH", "news.sports", "Game on!"]).await;
     
-    // In Redis, the subscriber gets the message twice (once for channel, once for pattern)
-    // But PUBLISH returns the number of clients, not messages
-    assert_eq!(result, RespValue::Integer(1)); // 1 client received it
+    // The subscriber gets the message twice (once for channel, once for pattern)
+    // PUBLISH returns 2 because both subscriptions matched (even though same client)
+    assert_eq!(result, RespValue::Integer(2));
 
     server.stop().await;
 }
